Move level3 tests to level2_test.go
